# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ROSA Regional Platform API
  description: Platform API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /clusters:
    get:
      summary: List user's clusters
      description: Retrieve clusters owned by the authenticated user
      operationId: listUserClusters
      tags:
        - Clusters
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of clusters to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of clusters to skip (pagination)
      responses:
        "200":
          description: List of user's clusters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create cluster
      description: |
        Create a new cluster for the authenticated user.

        The API automatically enriches the cluster spec with:
        - hostedCluster.issuerURL: Auto-populated from the regional OIDC CloudFront distribution
        - hostedCluster.release.image: Defaulted if not provided
        - hostedCluster.networking: Default CIDR ranges applied if not provided

        Possible error codes:
        - CLUSTERS-MGMT-CREATE-001: Invalid request body
        - CLUSTERS-MGMT-CREATE-002: Invalid cluster spec
        - CLUSTERS-MGMT-CREATE-003: Failed to create cluster
        - CLUSTERS-MGMT-CREATE-004: Failed to validate cluster name
        - CLUSTERS-MGMT-CREATE-005: Duplicate cluster name
      operationId: createUserCluster
      tags:
        - Clusters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterCreateRequest"
      responses:
        "201":
          description: Cluster created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /clusters/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
        description: Cluster ID
    get:
      summary: Get cluster details
      description: Retrieve cluster details (user must own the cluster)
      operationId: getUserCluster
      tags:
        - Clusters
      responses:
        "200":
          description: Cluster details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      summary: Update cluster
      description: Update cluster spec (user must own the cluster)
      operationId: patchUserCluster
      tags:
        - Clusters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterUpdateRequest"
      responses:
        "200":
          description: Cluster updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      summary: Update cluster
      description: Update cluster spec (user must own the cluster). Same behavior as PATCH.
      operationId: updateUserCluster
      tags:
        - Clusters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterUpdateRequest"
      responses:
        "200":
          description: Cluster updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cluster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete cluster
      description: Delete cluster (user must own the cluster)
      operationId: deleteUserCluster
      tags:
        - Clusters
      responses:
        "202":
          description: Cluster deletion initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  cluster_id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  schemas:
...
    Error:
      type: object
      description: Error response
      required:
        - kind
        - code
        - reason
      properties:
        kind:
          type: string
          description: Resource kind
          example: Error
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
...
    ClusterSpec:
      type: object
      description: Cluster specification defining desired state
      additionalProperties: true
      properties:
        deleteProtection:
          description: |-
            DeleteProtection prevents accidental deletion when enabled
            +hyperfleet:write-mode=mutable
          type: boolean
        displayName:
          description: |-
            DisplayName is a human-readable name for the cluster
            +hyperfleet:write-mode=mutable
            +kubebuilder:validation:MaxLength=256
          type: string
        expirationTimestamp:
          description: |-
            ExpirationTimestamp marks when this cluster should be automatically deleted
            +hyperfleet:write-mode=mutable
          type: object
        hostedCluster:
          description: |-
            HostedCluster contains the full HyperShift HostedCluster configuration
            All fields are generated from upstream and have safe defaults (hidden + service-set)
            until explicitly reviewed and exposed
            +kubebuilder:validation:Required
          $ref: '#/components/schemas/HostedClusterSpecPassthrough'
        properties:
          description: |-
            Properties are arbitrary key-value pairs for customer metadata
            +hyperfleet:write-mode=mutable
          type: object
          additionalProperties:
            type: string
        tags:
          description: |-
            Tags are customer-defined labels for organizational purposes
            This is a TechPreview feature
            +hyperfleet:write-mode=mutable
            +openshift:enable:FeatureGate=HyperFleetAutoScaling
          type: object
          additionalProperties:
            type: string
        cloudUrl:
          type: string
          readOnly: true
          description: CloudFront URL with cluster ID (auto-populated by server)
          example: https://doku78iof5s87.cloudfront.net/cluster-123
        placement:
          type: string
          description: Management cluster name (auto-populated if not provided)
          example: management-cluster-us-east-1
...
    Cluster:
      description: |-
        Cluster represents a HyperFleet managed OpenShift cluster
        +kubebuilder:object:root=true
        +kubebuilder:subresource:status
        +kubebuilder:resource:scope=Namespaced
      type: object
      required:
        - spec
      properties:
        spec:
          $ref: '#/components/schemas/ClusterSpec'
        status:
          $ref: '#/components/schemas/ClusterStatus'
    ClusterCreateRequest:
      type: object
      description: |
        Request body for creating a cluster.

        The API automatically enriches the cluster spec with:
        - hostedCluster.issuerURL: Auto-populated from the regional OIDC CloudFront distribution
        - hostedCluster.release.image: Defaulted if not provided
        - hostedCluster.networking: Default CIDR ranges applied if not provided
        - creatorARN: Set from the caller's IAM identity
      required:
        - name
        - spec
      properties:
        name:
          type: string
          pattern: "^[a-z0-9]([a-z0-9-]*[a-z0-9])?$"
          minLength: 3
          maxLength: 63
          description: DNS-compatible cluster name
        target_project_id:
          type: string
          description: Target project ID
        spec:
          allOf:
            - $ref: '#/components/schemas/ClusterSpec'
          description: Cluster specification (cloudUrl is server-populated and ignored on create)
    ClusterUpdateRequest:
      type: object
      description: Request body for updating a cluster
      required:
        - spec
      properties:
        spec:
          allOf:
            - $ref: '#/components/schemas/ClusterSpec'
          description: Cluster specification (mutable fields only)
...
    ClusterList:
      description: |-
        ClusterList contains a list of Clusters
        +kubebuilder:object:root=true
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            type: object
    # NodePool Schemas
...
    ClusterStatus:
      description: ClusterStatus defines the observed state of a Cluster
      type: object
      properties:
        apiEndpoint:
          description: APIEndpoint is the cluster API server endpoint
          type: string
        conditions:
          description: Conditions represent detailed cluster status
          type: array
          items:
            type: object
        consoleUrl:
          description: ConsoleURL is the web console URL
          type: string
        provisionStartTime:
          description: ProvisionStartTime is when provisioning began
          type: object
        readyTime:
          description: ReadyTime is when the cluster became ready
          type: object
        state:
          description: |-
            State represents the high-level cluster state
            +kubebuilder:validation:Enum=pending;provisioning;ready;degraded;deleting;failed
          type: string
        version:
          description: Version is the observed OpenShift version
          type: string
...
    HostedClusterSpecPassthrough:
      description: HostedClusterSpecPassthrough mirrors HostedClusterSpec from upstream HyperShift
      type: object
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"